+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 ## Change log
 ----------------------
 
+Version 5.1
+-------------
+
+CHANGED:
+
+- unit tests extended for improve code coverage to 100%
+- deleted unused method CalculateDateExtensions#calculateTimeFromNow
+- deleted unused method CalculateDateExtensions#calculateElapsedTimeInSeconds
+- deleted unused class TimeZoneExtensions
+
 Version 5
 -------------